streamlit_app/server_requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
numpy==2.1.3
pandas==2.2.3
faiss-cpu==1.9.0
joblib==1.4.2
scikit-learn==1.5.2
torch==2.3.0
tqdm==4.67.1
transformers==4.46.3
umap-learn==0.5.7
optuna==4.1.0
lightgbm==4.5.0
accelerate==1.1.1